move network_setup_and_teardown to conftest. Fix pylint complaints
